feat: Complete integration tests for RabbitMQ and MySQL
- Implemented the eceive component to consume messages from RabbitMQ and save them to MySQL.
- Added integration tests in eceive_test.go to verify the functionality.
- Configured Docker Compose to start RabbitMQ and MySQL services.
- Created migration files for setting up the database schema.
- Added logic to run migrations programmatically and via CLI.
- Updated README.md with setup and running instructions.

Test is completed and verified successfully.

+++ db/migrations/0001_create_payment_events_table.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE payment_events (
     id INT AUTO_INCREMENT PRIMARY KEY,
     user_id INT NOT NULL,
-    deposit_amount FLOAT NOT NULL,
+    deposit_amount INT NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
